+++ src/OpcUaModbusServer/Modbus/ModbusRTU/ModbusRTUClient.h
@@ -15,9 +15,32 @@
    Autor: Kai Huebl (kai@huebl-sgh.de)
  */
 
-#include "Modbus/Modbus/ModbusRTUClientTrx.h"
+#ifndef __Modbus_ModbusRTUClient_h__
+#define __Modbus_ModbusRTUClient_h__
+
+#include <stdint.h>
+#include <functional>
+
+#include "Modbus/ModbusRTU/ModbusRTU.h"
+#include "Modbus/ModbusRTU/ModbusRTUClientTrx.h"
 
 namespace Modbus
 {
 
+	class ModbusRTUClient
+	: public ModbusRTU
+	{
+	  public:
+		ModbusRTUClient(void);
+		virtual ~ModbusRTUClient(void);
+
+		bool readCoilReq(ReadCoilResFunc readCoilResFunc, uint8_t slave, uint16_t address, uint16_t numberCoils);
+
+	  private:
+
+		void handleReadCoilRes(const boost::system::error_code& ec, const ModbusRTUTrx::SPtr& modbusRTUTrx);
+	};
+
 }
+
+#endif
