+++ src/OpcUaModbusServer/Modbus/ModbusRTU/ModbusRTUServer.cpp
@@ -15,7 +15,7 @@
    Autor: Kai Huebl (kai@huebl-sgh.de)
  */
 
-#include "Modbus/Modbus/ModbusRTUServer.h"
+#include "Modbus/ModbusRTU/ModbusRTUServer.h"
 
 namespace Modbus
 {
